requirements.txt
asgiref==3.4.1
attrs==21.4.0
Django==4.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.3.3
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
PyJWT==2.3.0
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
validate-docbr==1.10.0
whitenoise==5.3.0
